+++ src/functions/project-dependencies-check-javascript.ts
@@ -1,0 +1,100 @@
+import { WorkspaceContext } from '@causa/workspace';
+import { ProjectDependenciesCheck } from '@causa/workspace-core';
+import type { VulnerabilityLevels } from 'audit-ci';
+import auditCi from 'audit-ci';
+import { TypeScriptConfiguration } from '../configurations/index.js';
+
+/**
+ * A valid vulnerability level of `audit-ci`.
+ */
+type VulnerabilityLevel = keyof VulnerabilityLevels;
+
+/**
+ * The default vulnerability level failing the dependencies check.
+ */
+const DEFAULT_VULNERABILITY_LEVEL: VulnerabilityLevel = 'low';
+
+/**
+ * The list of allowed vulnerability levels, defined by `audit-ci`.
+ */
+const ALLOWED_VULNERABILITY_LEVELS: VulnerabilityLevel[] = [
+  'low',
+  'moderate',
+  'high',
+  'critical',
+];
+
+/**
+ * Implements the {@link ProjectDependenciesCheck} function for JavaScript and TypeScript projects.
+ * This uses `audit-ci`, which is a wrapper around `npm audit`.
+ */
+export class ProjectDependenciesCheckForJavaScript extends ProjectDependenciesCheck {
+  async _call(context: WorkspaceContext): Promise<void> {
+    const conf = context.asConfiguration<TypeScriptConfiguration>();
+    const level = conf.get('javascript.dependenciesCheck.level');
+    const skipDev = conf.get('javascript.dependenciesCheck.skipDev');
+    const allowlist = conf.get('javascript.dependenciesCheck.allowlist');
+
+    if (level && !ALLOWED_VULNERABILITY_LEVELS.includes(level)) {
+      throw new Error(`Invalid dependencies check level '${level}'.`);
+    }
+
+    context.logger.info('🔍 Checking for vulnerable dependencies.');
+
+    await this.auditDependencies(context, { allowlist, skipDev, level });
+
+    context.logger.info('✅ No vulnerable dependency found.');
+  }
+
+  _supports(context: WorkspaceContext): boolean {
+    return ['javascript', 'typescript'].includes(
+      context.get('project.language') ?? '',
+    );
+  }
+
+  /**
+   * Audits the dependencies of the project using `audit-ci`.
+   * Handles the mapping of `audit-ci` options.
+   *
+   * @param context The workspace context.
+   * @param options Options when auditing dependencies.
+   */
+  private async auditDependencies(
+    context: WorkspaceContext,
+    options: {
+      allowlist?: string[];
+      skipDev?: boolean;
+      level?: VulnerabilityLevel;
+    } = {},
+  ): Promise<void> {
+    const projectPath = context.getProjectPathOrThrow();
+    const allowlist = options.allowlist ?? [];
+    const skipDev = options.skipDev ?? false;
+    const level = options.level ?? DEFAULT_VULNERABILITY_LEVEL;
+
+    const levels = auditCi.mapVulnerabilityLevelInput({ [level]: true });
+
+    await auditCi.npmAudit({
+      // Defaults obtained from the `audit-ci` source code.
+      // Some may not even be used (e.g. `package-manager`), as `npmAudit` is called directly.
+      // However they are all required.
+      'pass-enoaudit': false,
+      'retry-count': 5,
+      report: false,
+      'package-manager': 'npm',
+      'show-not-found': true,
+      'show-found': true,
+      registry: undefined,
+      summary: false,
+      'output-format': 'text',
+      'extra-args': [],
+      // Options actually set to a value.
+      directory: projectPath,
+      'report-type': 'summary',
+      allowlist: auditCi.Allowlist.mapConfigToAllowlist({ allowlist }),
+      ...levels,
+      levels,
+      'skip-dev': skipDev,
+    });
+  }
+}