+++ src/functions/project-push-artefact-npm-package.spec.ts
@@ -112,7 +112,7 @@ describe('ProjectPushArtefactForNpmPackage', () => {
     });
 
     expect(actualDestination).toEqual(tmpDir);
-    expect(npmService.publish).toHaveBeenCalledWith({
+    expect(npmService.publish).toHaveBeenCalledExactlyOnceWith({
       workingDirectory: tmpDir,
     });
   });